chore(server): move properties to compare to comparators
