+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/flattened-index-field-metadata.ts
@@ -1,0 +1,15 @@
+import { IndexFieldMetadataEntity } from 'src/engine/metadata-modules/index-metadata/index-field-metadata.entity';
+
+const indexFieldMetadataRelationProperties = [
+  'indexMetadata',
+  'fieldMetadata',
+] as const satisfies (keyof IndexFieldMetadataEntity)[];
+
+type IndexFieldMetadataRelationProperties =
+  (typeof indexFieldMetadataRelationProperties)[number];
+
+export type FlattenedIndexFieldMetadata = Partial<
+  Omit<IndexFieldMetadataEntity, IndexFieldMetadataRelationProperties>
+> & {
+  uniqueIdentifier: string;
+};
