+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-field-input.ts
@@ -1,28 +1,7 @@
 import { FieldMetadataEntity } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
 
-export const fieldMetadataEntityEditableProperties = [
-  'defaultValue',
-  'description',
-  'icon',
-  'isActive',
-  'isLabelSyncedWithName',
-  'isUnique',
-  'label',
-  'name',
-  'options',
-  'standardOverrides',
-] as const satisfies (keyof WorkspaceMigrationFieldInput)[];
-export type FieldMetadataEntityEditableProperties =
-  (typeof fieldMetadataEntityEditableProperties)[number];
-
-// TODO could describe required minimum keys
 export type WorkspaceMigrationFieldInput = Partial<
   Omit<FieldMetadataEntity, 'object' | 'indexFieldMetadatas'>
 > & {
   uniqueIdentifier: string;
 };
-
-export const fieldMetadataPropertiesToStringify = [
-  'defaultValue',
-  'standardOverrides',
-] as const satisfies FieldMetadataEntityEditableProperties[];
