--- packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-index-input.ts
@@ -1,33 +1,0 @@
-import { IndexFieldMetadataEntity } from 'src/engine/metadata-modules/index-metadata/index-field-metadata.entity';
-import { IndexMetadataEntity } from 'src/engine/metadata-modules/index-metadata/index-metadata.entity';
-
-export const indexMetadataEntityPropertiesToCompare = [
-  'flattenedIndexFieldMetadatas', // Comparing this as whole ? should iterate on each keys ? => TBD should only map over cols as before ?
-  'indexType',
-  'indexWhereClause',
-  'isUnique',
-  'name',
-] as const satisfies (keyof FlattenedIndexMetadata)[];
-export type IndexMetadataEntityEditableProperties =
-  (typeof indexMetadataEntityPropertiesToCompare)[number];
-
-const indexFieldMetadataRelationProperties = [
-  'indexMetadata',
-  'fieldMetadata',
-] as const satisfies (keyof IndexFieldMetadataEntity)[];
-
-type IndexFieldMetadataRelationProperties =
-  (typeof indexFieldMetadataRelationProperties)[number];
-
-export type FlattenedIndexFieldMetadata = Partial<
-  Omit<IndexFieldMetadataEntity, IndexFieldMetadataRelationProperties>
-> & {
-  uniqueIdentifier: string;
-};
-
-export type FlattenedIndexMetadata = Partial<
-  Omit<IndexMetadataEntity, 'indexFieldMetadatas' | 'objectMetadata'> // Might have an issue as ObjectMetadataId != uniqueIdentifier
-> & {
-  flattenedIndexFieldMetadatas: FlattenedIndexFieldMetadata[];
-  uniqueIdentifier: string;
-};
