+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-action-v2.ts
@@ -1,9 +1,7 @@
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import { FromTo } from 'src/engine/workspace-manager/workspace-migration-v2/types/from-to.type';
-import {
-  ObjectMetadataEntityEditableProperties,
-  WorkspaceMigrationObjectWithoutFields,
-} from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-input';
+import { WorkspaceMigrationObjectWithoutFields } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-input';
+import { WorkspaceMigrationObjectInputPropertiesToCompare } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/workspace-migration-object-metadata-input-comparator.util';
 
 type ObjectActionCommon = {
   objectMetadataInput: WorkspaceMigrationObjectWithoutFields;
@@ -16,10 +14,10 @@ export type UpdateObjectAction = {
   type: 'update_object';
   updates: Partial<
     {
-      [P in ObjectMetadataEntityEditableProperties]: {
+      [P in WorkspaceMigrationObjectInputPropertiesToCompare]: {
         property: P;
       } & FromTo<ObjectMetadataEntity[P]>;
-    }[ObjectMetadataEntityEditableProperties]
+    }[WorkspaceMigrationObjectInputPropertiesToCompare]
   >[];
 } & ObjectActionCommon;
