+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/get-workspace-migration-v2-index-actions.ts
@@ -1,8 +1,8 @@
+import { FlattenedIndexMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/types/flattened-index-metadata';
 import {
   CreateIndexAction,
   DeleteIndexAction,
 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-index-action-v2';
-import { FlattenedIndexMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-index-input';
 
 export const getWorkspaceMigrationV2CreateIndexAction = (
   flattenedIndexMetadata: FlattenedIndexMetadata,
